backend/requirements.txt
# backend/requirements.txt
tensorflow
jax
jaxlib
numpy
dm-haiku
matplotlib
opencv-python
scikit-learn
pandas
flask
tqdm
flask-cors==6.0.0
gunicorn
dotenv
supabase
PyJWT #token system for the authentication
python-dotenv